requirements.txt
streamlit
pandas
mysql-connector-python
google-generativeai
python-dotenv